refactor: extract auth API calls into dedicated functions and update store
